internal: drop 1.21.10 support

+++ src/main/kotlin/xyz/aerii/athen/modules/impl/render/radial/impl/RadialMenu.kt
@@ -431,10 +431,10 @@ object RadialMenu : Module(
             val s = i * step + gap * .5f + off
             val e = (i + 1) * step - gap * .5f + off
 
-            val c0 = Mth.cos(s/*? >= 1.21.11 {*//*.toDouble()*//*? }*/)
-            val s0 = Mth.sin(s/*? >= 1.21.11 {*//*.toDouble()*//*? }*/)
-            val c1 = Mth.cos(e/*? >= 1.21.11 {*//*.toDouble()*//*? }*/)
-            val s1 = Mth.sin(e/*? >= 1.21.11 {*//*.toDouble()*//*? }*/)
+            val c0 = Mth.cos(s.toDouble())
+            val s0 = Mth.sin(s.toDouble())
+            val c1 = Mth.cos(e.toDouble())
+            val s1 = Mth.sin(e.toDouble())
 
             val x0 = cx + outer * c0
             val y0 = cy + outer * s0
